requirements.txt
openai>=1.52.0
pandas>=2.2.2
pydantic>=2.9.0
python-dotenv>=1.0.1
PyYAML>=6.0.1
pdfplumber>=0.11.0
python-docx>=1.1.2
tqdm>=4.66.5
fastapi>=0.115.0
uvicorn>=0.30.1
python-multipart>=0.0.9
